requirements.txt
numpy==1.16.5
scipy==1.2.1
pandas==0.24.2
opencv-python==4.2.0.32
matplotlib==2.2.3
scikit-learn==0.20.4
scikit-image==0.14.2
Pillow==6.2.1
PyYAML==5.2
h5py==2.10.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
tensorflow-gpu==1.12.0
protobuf==3.10.0
progressbar==2.5